shift calculateInputs to core
prepare to deprecate ./create
