+++ src/transactions/core.js
@@ -1,7 +1,59 @@
 import { num2VarInt, num2hexstring, StringStream, reverseHex, hash256 } from '../utils'
-import { generateSignature, getVerificationScriptFromPublicKey, getPublicKeyFromPrivateKey } from '../wallet'
+import { generateSignature, getVerificationScriptFromPublicKey, getPublicKeyFromPrivateKey, getScriptHashFromAddress } from '../wallet'
 import { serialize as serialiseExclusive, deserialize as deserializeExclusive } from './exclusive'
+import { ASSETS, ASSET_ID } from '../consts'
 import * as comp from './components'
+
+/**
+ * Calculate the inputs required given the intents and gasCost. gasCost has to be seperate because it will not be reflected as an TransactionOutput.
+ * @param {Balance} balances - Balance of all assets available.
+ * @param {TransactionOutput[]} intents - All sending intents
+ * @param {number} gasCost - gasCost required for the transaction.
+ * @return {Object} {inputs: TransactionInput[], change: TransactionOutput[] }
+ */
+export const calculateInputs = (balances, intents, gasCost = 0) => {
+  // We will work in integers here to be more accurate.
+  // As assets are stored as Fixed8, we just multiple everything by 10e8 and round off to get integers.
+  const requiredAssets = intents.reduce((assets, intent) => {
+    const fixed8Value = Math.round(intent.value * 100000000)
+    assets[intent.assetId] ? assets[intent.assetId] += fixed8Value : assets[intent.assetId] = fixed8Value
+    return assets
+  }, {})
+  // Add GAS cost in
+  if (gasCost > 0) {
+    const fixed8GasCost = gasCost * 100000000
+    requiredAssets[ASSET_ID.GAS] ? requiredAssets[ASSET_ID.GAS] += fixed8GasCost : requiredAssets[ASSET_ID.GAS] = fixed8GasCost
+  }
+  let change = []
+  const inputs = Object.keys(requiredAssets).map((assetId) => {
+    const requiredAmt = requiredAssets[assetId]
+    const assetBalance = balances[ASSETS[assetId]]
+    if (assetBalance.balance * 100000000 < requiredAmt) throw new Error(`Insufficient ${ASSETS[assetId]}! Need ${requiredAmt / 100000000} but only found ${assetBalance.balance}`)
+    // Ascending order sort
+    assetBalance.unspent.sort((a, b) => a.value - b.value)
+    let selectedInputs = 0
+    let selectedAmt = 0
+    // Selected min inputs to satisfy outputs
+    while (selectedAmt < requiredAmt) {
+      selectedInputs += 1
+      if (selectedInputs > assetBalance.unspent.length) throw new Error(`Insufficient ${ASSETS[assetId]}! Reached end of unspent coins!`)
+      selectedAmt += Math.round(assetBalance.unspent[selectedInputs - 1].value * 100000000)
+    }
+    // Construct change output
+    if (selectedAmt > requiredAmt) {
+      change.push({
+        assetId,
+        value: (selectedAmt - requiredAmt) / 100000000,
+        scriptHash: getScriptHashFromAddress(balances.address)
+      })
+    }
+    // Format inputs
+    return assetBalance.unspent.slice(0, selectedInputs).map((input) => {
+      return { prevHash: input.txid, prevIndex: input.index }
+    })
+  }).reduce((prev, curr) => prev.concat(curr), [])
+  return { inputs, change }
+}
 
 /**
  * Serializes a given transaction object
